idk
- 9Khasaanda(Horc) ~ line 46-48 is broke the order doesnt matter yet it only does the first kill & the map item, it leaves the 2nd killQuest not done and fails to complete.

EnoughDOOMforanArchfiend ----
WillpowerExtraction /
NulgathDemandsWork somewhere is broke somewhere its only kill these monsters once then failing to bank, stopping the script gives "Error in timer thread: The operation was canceled." in the debugger. ive looked though the scripts and ive found nothing outa the ordinary to my eyes...

mobs so far;
Tainted Elemental,
Laken... are the ones so far ive changed some to Huntmonster but :cry:  idt itll work... this is for tomarrow as exe is done for today and mah brains to fried to do this properly.