Another fix for CheckInventory
ItemName and ItemID fixed - stringArray works when its called with items in the inventory - if its in the bank it wont - no clue how to fix it. The fix is messy, but it works. Pretty sure I can put the BankCheck in ItemName before the inv check to make it cleaner, but that would spam the bank each time so dunno if you want that. This is the extent of how I can fix it until someone with a brain looks at it x)