Added full Apotheosis merge and misc fixes
Replaced while loops with killmonsters and some other stuff u guys recommended. Added full apotheosis merge too. Stops if missing the required total number of insignias though. Maybe a try/catch would be better for the apotheosis merging part so that if user has some of the weapons already they won't have to get extra insignias for it to continue but idk how to implement that well.