Update CoreSDKA.cs
sorry i haven't checked this for a bit. 

So the issue for me was there was no check for the case of having the accursed arsenic without the CSO's and Ominous Auras, in that it would jump straight from line 340 to line 359 to buy the accursed arsenic of doom. not sure how I made it past the checks before hand to get the accursed arsenic becos I didn't farm for it myself with the script, but I included it anyways