logic issue fix
this solves the issue if the user already has the armor but doesn't have an active membership. since the SDKA items only requires the armor and not an active membership
hence the && checking so if the user has the armor and is not a member
which the code looks like this
if( false && true ) and this makes the statement false.