Cockroach Moe
Hello, I'm back again.

The Cockroach Moe bot farms almost all Endgame items and Ultra Boss requisites, with an option to farm ArchMage alongside. Last time, the main issue with the bot was that it copy pasted code from other scripts to circumvent the bot stopping because of dailies. The issue has been fixed by checking if the daily items are present, or can be obtained, before committing to the farm. Aside from a few extra functions for clearing clutter, everything should be structured and commented on well enough.

This time I'm more committed to getting this bot on Skua, so please let me know what needs to be fixed.